changed Question Names and extracted the header and button from app
