+++ src/components/SubmitButton.jsx
@@ -1,0 +1,9 @@
+import "./SubmitButton.css";
+
+export const SubmitButton = () => {
+  return (
+    <button type="submit" className="submit-button">
+      Submit your answer
+    </button>
+  );
+};
